roles done done gender changed users started jquery added login started

+++ backoffice/content/gender/delete.php
@@ -3,8 +3,6 @@ if ($_SERVER["REQUEST_METHOD"] === "GET") {
     $id = isset($_GET['id']) ? mysqli_real_escape_string($conn, $_GET['id']) : '';
     $operacao = isset($_GET['operation']) ? mysqli_real_escape_string($conn, $_GET['operation']) : '';
 
-    var_dump($id, $operacao); // Add this line for debugging purposes
-
     if (empty($id) || $operacao !== "eliminar") {
         echo "Erro, pedido inválido";
         exit();
@@ -20,12 +18,28 @@ $resultado = mysqli_query($conn, $sql);
 if (!$resultado) {
     echo 'Falha na consulta: ' . mysqli_error($conn);
 } else {
-    $sql = "DELETE FROM gender WHERE id ='$id'";
-    if (!mysqli_query($conn, $sql)) {
-        echo "ERRO - Falha ao executar o comando: \"$sql\" <br>" . mysqli_error($conn);
+
+
+    $checkuser = "SELECT * FROM user WHERE gender = '$id'";
+    $resuser = mysqli_query($conn, $checkuser);
+    if ($resuser && mysqli_num_rows($resuser) == 0) {
+        $sql = "DELETE FROM gender WHERE id ='$id'";
+        if (!mysqli_query($conn, $sql)) {
+            echo "ERRO - Falha ao executar o comando: \"$sql\" <br>" . mysqli_error($conn);
+        } else {
+            header('Location: ./?p=6');
+            exit();
+        }
     } else {
-        header('Location: ./?p=6');
-        exit();
+        $sql = "SELECT name FROM user WHERE gender = '$id'";
+
+        $showuser = mysqli_query($conn, $sql);
+
+        echo "Género necessário na tabela User.<br>";
+        while ($row = mysqli_fetch_array($showuser)) {
+            $name = $row['name'];
+            echo "Nome: \"$name\" <br>" . mysqli_error($conn);
+        }
     }
 }
 ?>
